+++ x-pack/plugins/cases/server/client/cases/create.ts
@@ -27,7 +27,7 @@ import {
   throwIfCustomFieldTypesInvalid,
   throwIfDuplicatedCustomFieldKeysInRequest,
   throwIfMissingRequiredCustomField,
-} from './validation';
+} from './validators';
 
 async function validateCustomFieldsInRequest({
   data,
